pubsub_writer/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==2.18.*
